The texture to use for terrain is determined in the fragment shader by pulling predetermined values from a 1D texture

+++ Src/World/TerrainData.h
@@ -19,7 +19,7 @@ namespace t3d
 		//TODO PURE HACKERY HERE
 		HeightMap& heightMap() { return mHeightMap; }
 
-		typedef std::vector<GLint> TextureIndicies;
+		typedef std::vector<GLubyte> TextureIndicies;
 		TextureIndicies& textureIndicies() { return mTextureIndicies; }
 
 	private:
